Implement signal handling for Unix
This is done using the "self-pipe" trick. The Tokio-based driver waits
for readability on the read end of the pipe until the signal handler
writes a single byte to the write end. This allows for communicating
with the outside from within the primitive world of the signal handler.

The current implementation fulfills futures only once, with subsequent
polls returning `Ready`. This is done by enabling the signal's `caught`
bit. This saves having to check readability of the pipe again.

TODO items:

* Implement futures for Windows.

* Implement futures for reasonably reusable signals.

* Check whether a signal has already been registered. Alternatively, we
  can instead allow for registering multiple signal handlers by queuing
  their wakers. Although this seems rather odd and perhaps unnecessary.